REACT : Add smp_auth custom oauth login

+++ src/components/common/auth/login.jsx
@@ -4,9 +4,12 @@ import styles from './login.module.css';
 import Logo from '../logo/logo';
 import Goback from '../goback/goback';
 
-const Login = ({ authService, smpChat }) => {
+const Login = ({ authService, smpAuth, smpChat }) => {
   const history = useHistory();
   const auth = localStorage.getItem('state');
+  const params = new URLSearchParams(window.location.search);
+  const code = params.get('code');
+  const state = params.get('state');
 
   const goToStudy = useCallback(
     (id) => {
@@ -18,33 +21,58 @@ const Login = ({ authService, smpChat }) => {
     [history]
   );
 
-  const onClickLoginButton = (e) => {
-    let loginType = e.target.dataset.name;
+  const onClickLoginButton = useCallback(
+    (e) => {
+      let loginType = e.target.dataset.name;
 
-    if (!loginType) {
-      const li = e.target.closest('li');
-      loginType = li.dataset.name;
+      if (!loginType) {
+        const li = e.target.closest('li');
+        loginType = li.dataset.name;
+      }
+
+      if (loginType === 'Smpark') {
+        smpAuth.login();
+      } else {
+        authService
+          .login(loginType)
+          .then((data) => {
+            localStorage.setItem('state', data.user.uid);
+
+            smpChat.load(data.user.uid);
+
+            goToStudy(data.user.uid);
+          })
+          .catch((e) => {
+            e.code === 'auth/account-exists-with-different-credential' &&
+              alert(
+                `같은 이메일 주소가 등록되어 있습니다. 기존의 등록한 방식으로 로그인하여 주십시오. ${e.email}`
+              );
+
+            localStorage.clear();
+          });
+      }
+    },
+    [authService, goToStudy, smpAuth, smpChat]
+  );
+
+  useEffect(() => {
+    if (code && state) {
+      smpAuth
+        .token(code, state)
+        .then((result) => {
+          authService.customLogin(result.token, (user) => {
+            localStorage.setItem('state', user.uid);
+
+            window.close();
+          });
+        })
+        .catch((e) => {
+          localStorage.clear();
+          window.close();
+          throw Error(e);
+        });
     }
-
-    authService
-      .login(loginType)
-      .then((data) => {
-        localStorage.setItem('state', data.user.uid);
-
-        smpChat.load(data.user.uid);
-
-        goToStudy(data.user.uid);
-      })
-      .catch((e) => {
-        console.log(e);
-        e.code === 'auth/account-exists-with-different-credential' &&
-          alert(
-            `같은 이메일 주소가 등록되어 있습니다. 기존의 등록한 방식으로 로그인하여 주십시오. ${e.email}`
-          );
-
-        localStorage.clear();
-      });
-  };
+  }, [smpAuth, code, state, goToStudy, authService, smpChat]);
 
   useEffect(() => {
     authService.loginUserCheck((user) => {
@@ -52,7 +80,6 @@ const Login = ({ authService, smpChat }) => {
         if (!auth) {
           localStorage.setItem('state', user.uid);
         }
-
         goToStudy(user.uid);
       } else {
         if (auth) {
@@ -66,7 +93,7 @@ const Login = ({ authService, smpChat }) => {
 
   return (
     <>
-      {!auth && (
+      {!auth && !code && !state && (
         <section className={styles.section}>
           <header className={styles.header}>
             <Goback backBox={styles.backBox} />
